Added confirmation before creating classes and lectures

+++ study.py
@@ -21,33 +21,43 @@ def main(argv):
 
     location = directory + args[0] + "/" + args[1]
     if not os.path.exists(directory + args[0]):
-        os.mkdir(directory + args[0])
-    else:
-        print("Directory '" + args[0] + "' already exists")
+        print("Course '" + args[0] + "' does not exist yet. Create it? [Y/n]")
+        create = input()
+        if not (create == "n" or create == "N"):
+            os.mkdir(directory + args[0])
+        else:
+            sys.exit(0)
 
     if not os.path.exists(location):
-        os.mkdir(location);
+        print ("Lecture '" +  args[1]  + "' does not exist yet. Create it? [Y/n]")
+        create = input()
+        if not (create == "n" or create == "N"):
+            os.mkdir(location);
+            print("Title of Lecture: ")
+            title = input()
+
+            texFileLocation = location + "/" + args[1] + ".tex"
+
+            f = open(texFileLocation, "w")
+            f.write("\input{preamble}\n")
+            f.write("\\title{" + title + "}")
+            f.write("\\begin{document}\n")
+            f.write("\\maketitle\n\n")
+            f.write("\\end{document}\n")
+            f.close()
+            copyfile("./preamble.tex", location + "/preamble.tex")
+
+            os.chdir(location);
+            os.system('pdflatex -output-directory '+ location + ' ' + texFileLocation)
+            os.system('evince ' + location + "/" + args[1] + ".pdf" + '&')
+            os.system('vim ' + texFileLocation)
+        else:
+            sys.exit(0);
     else:
-        print("Directory '" + args[1] + "' already exists")
-
-    print("Title of Lecture: ")
-    title = input()
-
-    texFileLocation = location + "/" + args[1] + ".tex"
-
-    f = open(texFileLocation, "w")
-    f.write("\input{preamble}\n")
-    f.write("\\title{" + title + "}")
-    f.write("\\begin{document}\n")
-    f.write("\\maketitle\n\n")
-    f.write("\\end{document}\n")
-    f.close()
-    copyfile("./preamble.tex", location + "/preamble.tex")
-
-    os.chdir(location);
-    os.system('pdflatex -output-directory '+ location + ' ' + texFileLocation)
-    os.system('evince ' + location + "/" + args[1] + ".pdf" + '&')
-    os.system('vim ' + texFileLocation)
+        texFileLocation = location + "/" + args[1] + ".tex"
+        os.chdir(location);
+        os.system('evince ' + location + "/" + args[1] + ".pdf" + '&')
+        os.system('vim ' + texFileLocation)
 
 if __name__ == "__main__":
     main(sys.argv[1:])
